+++ anagrafica/models.py
@@ -676,7 +676,9 @@ class Persona(ModelloSemplice, ConMarcaTemporale, ConAllegati, ConVecchioID):
 
     def richieste_di_partecipazione(self, corso_stato=None):
         """ Restituisce richieste di partecipazione ai corsi confermate e in attesa """
+
         from formazione.models import PartecipazioneCorsoBase, CorsoBase
+
         if corso_stato is None:
             corso_stato = [CorsoBase.ATTIVO,]
 
@@ -685,6 +687,14 @@ class Persona(ModelloSemplice, ConMarcaTemporale, ConAllegati, ConVecchioID):
         requests_to_courses = confirmed | pending
 
         return requests_to_courses.filter(persona=self, corso__stato__in=corso_stato)
+
+    @property
+    def corsi(self):
+        from formazione.models import CorsoBase
+
+        richieste_di_partecipazione = self.richieste_di_partecipazione().values_list('corso_id', flat=True)
+        corsi_in_attesa_o_confermati = CorsoBase.objects.filter(id__in=richieste_di_partecipazione)
+        return corsi_in_attesa_o_confermati | self.corsi_frequentati
 
     @property
     def corsi_frequentati(self):
@@ -2353,6 +2363,12 @@ class Delega(ModelloSemplice, ConStorico, ConMarcaTemporale):
             delega.termina(mittente=mittente, accoda=True, termina_at=datetime.now())
 
         return numero_deleghe
+
+    @classmethod
+    def corsi(cls, persona):
+        from formazione.models import CorsoBase
+        deleghe_direttore_corso = cls.objects.filter(persona=persona, tipo=DIRETTORE_CORSO)
+        return CorsoBase.objects.filter(pk__in=deleghe_direttore_corso.values_list('oggetto_id', flat=True))
 
 
 class Fototessera(ModelloSemplice, ConAutorizzazioni, ConMarcaTemporale):
