+++ anagrafica/urls/profilo.py
@@ -1,6 +1,7 @@
 from django.conf.urls import url
-from anagrafica import viste
 
+from anagrafica import viste
+from anagrafica.profile.views import profilo
 
 app_label = 'profilo'
 urlpatterns = [
@@ -10,6 +11,7 @@ urlpatterns = [
     url(r'^(?P<pk>[0-9]+)/documenti/(?P<documento_pk>[0-9]+)/cancella/$', viste.profilo_documenti_cancella),
     url(r'^(?P<pk>[0-9]+)/curriculum/(?P<tp_pk>[0-9]+)/cancella/$', viste.profilo_curriculum_cancella),
     url(r'^(?P<pk>[0-9]+)/sangue/(?P<donazione_pk>[0-9]+)/cancella/$', viste.profilo_sangue_cancella),
-    url(r'^(?P<pk>[0-9]+)/(?P<sezione>.*)/$', viste.profilo, name='profilo'),
-    url(r'^(?P<pk>[0-9]+)/$', viste.profilo, name="main"),
+
+    url(r'^(?P<pk>[0-9]+)/(?P<sezione>.*)/$', profilo, name='profilo'),
+    url(r'^(?P<pk>[0-9]+)/$', profilo, name="main"),
 ]
